코드 일부 수정/jjy

+++ coding-test/src/jjy/프로그래머스/x만큼_간격이_있는_n개의_숫자/Solution.java
@@ -5,7 +5,7 @@ class Solution {
         long[] answer = new long[n];
 
         for (int i = 0; i < n; i++) {
-            answer[i] = (i + 1) * x;
+            answer[i] = (long) (i + 1) * x;
         }
 
         return answer;
